pyproject.toml
[tool.poetry]
name = "herz-des-nordens"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psutil = "^5.9.8"
pytz = "^2023.3.post1"
googletrans = ">=3.1.0a0,<3.2.0"
aiohttp = ">=3.8.1,<3.9.0"
aiosignal = "^1.3.1"
requests = "^2.31.0"
py-dactyl = "^2.0.4"
speedtest-cli = "^2.1.3"
import-file = "^1.11"
py-cord = "^2.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
